split xyk tests into separate files
