+++ pallets/xyk/src/tests/mod.rs
@@ -1,3 +1,6 @@
+mod creation;
+mod fees;
 mod invariants;
+mod liquidity;
 pub(crate) mod mock;
-mod tests;
+mod trades;
