+++ pallets/xyk/src/tests/trades.rs
@@ -1,664 +1,14 @@
-// This file is part of Basilisk-node.
-
-// Copyright (C) 2020-2021  Intergalactic, Limited (GIB).
-// SPDX-License-Identifier: Apache-2.0
-
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-pub use super::mock::{
-	Currency, Event as TestEvent, ExtBuilder, Origin, System, Test, ACA, ALICE, BOB, CHARLIE, DOT, HDX,
-	HDX_DOT_POOL_ID, ONE, XYK,
-};
-use crate::*;
+pub use super::mock::*;
+use crate::{Error, Event};
 use frame_support::BoundedVec;
 use frame_support::{assert_noop, assert_ok};
-use hydra_dx_math::MathError;
 use hydradx_traits::AMM as AmmPool;
+use orml_traits::MultiCurrency;
 use pallet_asset_registry::AssetType;
 use sp_std::convert::TryInto;
 
-pub fn new_test_ext() -> sp_io::TestExternalities {
-	let mut ext = ExtBuilder::default().build();
-	ext.execute_with(|| System::set_block_number(1));
-	ext
-}
-
-fn expect_events(e: Vec<TestEvent>) {
-	e.into_iter().for_each(frame_system::Pallet::<Test>::assert_has_event);
-}
-
-#[test]
-fn create_pool_should_work() {
-	new_test_ext().execute_with(|| {
-		let asset_a = HDX;
-		let asset_b = ACA;
-		assert_ok!(XYK::create_pool(
-			Origin::signed(ALICE),
-			asset_a,
-			asset_b,
-			100_000_000_000_000,
-			Price::from(10)
-		));
-
-		let pair_account = XYK::get_pair_id(AssetPair {
-			asset_in: asset_a,
-			asset_out: asset_b,
-		});
-		let share_token = XYK::share_token(pair_account);
-
-		assert_eq!(XYK::get_pool_assets(&pair_account), Some(vec![asset_a, asset_b]));
-
-		assert_eq!(Currency::free_balance(asset_a, &pair_account), 100000000000000);
-		assert_eq!(Currency::free_balance(asset_b, &pair_account), 1000000000000000);
-		assert_eq!(Currency::free_balance(asset_a, &ALICE), 900000000000000);
-		assert_eq!(Currency::free_balance(asset_b, &ALICE), 0);
-		assert_eq!(Currency::free_balance(share_token, &ALICE), 100000000000000);
-		assert_eq!(XYK::total_liquidity(&pair_account), 100000000000000);
-
-		expect_events(vec![Event::PoolCreated {
-			who: ALICE,
-			asset_a,
-			asset_b,
-			initial_shares_amount: 100000000000000,
-			share_token,
-			pool: pair_account,
-		}
-		.into()]);
-	});
-}
-
-#[test]
-fn create_same_pool_should_not_work() {
-	new_test_ext().execute_with(|| {
-		let user = ALICE;
-		let asset_a = HDX;
-		let asset_b = ACA;
-
-		assert_ok!(XYK::create_pool(
-			Origin::signed(user),
-			asset_b,
-			asset_a,
-			1000,
-			Price::from(2)
-		));
-		assert_noop!(
-			XYK::create_pool(Origin::signed(user), asset_b, asset_a, 999, Price::from(2)),
-			Error::<Test>::InsufficientLiquidity
-		);
-		assert_noop!(
-			XYK::create_pool(Origin::signed(user), asset_b, asset_a, 1000, Price::from(0)),
-			Error::<Test>::ZeroInitialPrice
-		);
-		assert_noop!(
-			XYK::create_pool(Origin::signed(user), asset_a, asset_a, 1000, Price::from(2)),
-			Error::<Test>::CannotCreatePoolWithSameAssets
-		);
-		assert_noop!(
-			XYK::create_pool(Origin::signed(user), asset_b, asset_a, 1000, Price::from(2)),
-			Error::<Test>::TokenPoolAlreadyExists
-		);
-
-		let pair_account = XYK::get_pair_id(AssetPair {
-			asset_in: asset_a,
-			asset_out: asset_b,
-		});
-		let share_token = XYK::share_token(pair_account);
-
-		expect_events(vec![Event::PoolCreated {
-			who: ALICE,
-			asset_a: asset_b,
-			asset_b: asset_a,
-			initial_shares_amount: 2000,
-			share_token,
-			pool: pair_account,
-		}
-		.into()]);
-	});
-}
-
-#[test]
-fn create_pool_overflowing_amount_should_not_work() {
-	new_test_ext().execute_with(|| {
-		let user = ALICE;
-		let asset_a = HDX;
-		let asset_b = ACA;
-
-		assert_noop!(
-			XYK::create_pool(
-				Origin::signed(user),
-				asset_b,
-				asset_a,
-				u128::MAX as u128,
-				Price::from(2)
-			),
-			Error::<Test>::CreatePoolAssetAmountInvalid
-		);
-	});
-}
-
-#[test]
-fn create_pool_with_insufficient_balance_should_not_work() {
-	new_test_ext().execute_with(|| {
-		let user = ALICE;
-		let asset_a = HDX;
-
-		assert_noop!(
-			XYK::create_pool(
-				Origin::signed(user),
-				4000,
-				asset_a,
-				100_000_000_000_000,
-				Price::from(10)
-			),
-			Error::<Test>::InsufficientAssetBalance
-		);
-
-		assert_noop!(
-			XYK::create_pool(
-				Origin::signed(user),
-				asset_a,
-				4000,
-				100_000_000_000_000,
-				Price::from(10)
-			),
-			Error::<Test>::InsufficientAssetBalance
-		);
-	});
-}
-
-#[test]
-fn add_liquidity_should_work() {
-	new_test_ext().execute_with(|| {
-		let user = ALICE;
-		let asset_a = DOT;
-		let asset_b = HDX;
-
-		assert_ok!(XYK::create_pool(
-			Origin::signed(user),
-			asset_a,
-			asset_b,
-			100_000_000,
-			Price::from_float(0.654)
-		));
-		let pair_account = XYK::get_pair_id(AssetPair {
-			asset_in: asset_a,
-			asset_out: asset_b,
-		});
-		let share_token = XYK::share_token(pair_account);
-
-		assert_eq!(Currency::free_balance(asset_b, &pair_account), 65_400_000);
-		assert_eq!(XYK::total_liquidity(&pair_account), 65400000);
-
-		assert_ok!(XYK::add_liquidity(
-			Origin::signed(user),
-			asset_a,
-			asset_b,
-			400_000,
-			1_000_000_000_000
-		));
-
-		assert_eq!(Currency::free_balance(share_token, &user), 65661600);
-
-		assert_eq!(Currency::free_balance(asset_b, &pair_account), 65_661_601);
-		assert_eq!(Currency::free_balance(asset_a, &pair_account), 100400000);
-		assert_eq!(Currency::free_balance(asset_a, &user), 999999899600000);
-		assert_eq!(XYK::total_liquidity(&pair_account), 65661600);
-
-		expect_events(vec![
-			Event::PoolCreated {
-				who: ALICE,
-				asset_a,
-				asset_b,
-				initial_shares_amount: 65400000,
-				share_token,
-				pool: pair_account,
-			}
-			.into(),
-			Event::LiquidityAdded {
-				who: ALICE,
-				asset_a,
-				asset_b,
-				amount_a: 400000,
-				amount_b: 261601,
-			}
-			.into(),
-		]);
-	});
-}
-
-#[test]
-fn add_liquidity_mints_correct_shares() {
-	new_test_ext().execute_with(|| {
-		let user = ALICE;
-		let asset_a = DOT;
-		let asset_b = HDX;
-
-		assert_ok!(XYK::create_pool(
-			Origin::signed(user),
-			asset_a,
-			asset_b,
-			100_000_000,
-			Price::from_float(0.654)
-		));
-
-		assert_ok!(XYK::add_liquidity(
-			Origin::signed(user),
-			asset_b,
-			asset_a,
-			261600,
-			1_000_000_000_000
-		));
-
-		let pair_account = XYK::get_pair_id(AssetPair {
-			asset_in: asset_a,
-			asset_out: asset_b,
-		});
-		let share_token = XYK::share_token(pair_account);
-
-		assert_eq!(Currency::free_balance(share_token, &user), 65661600);
-	});
-}
-
-#[test]
-fn add_liquidity_as_another_user_should_work() {
-	new_test_ext().execute_with(|| {
-		let user = ALICE;
-		let asset_a = HDX;
-		let asset_b = ACA;
-		assert_ok!(XYK::create_pool(
-			Origin::signed(user),
-			asset_b,
-			asset_a,
-			100_000_000,
-			Price::from(10_000)
-		));
-		assert_ok!(XYK::add_liquidity(
-			Origin::signed(user),
-			asset_b,
-			asset_a,
-			400_000,
-			1_000_000_000_000
-		));
-
-		let pair_account = XYK::get_pair_id(AssetPair {
-			asset_in: asset_a,
-			asset_out: asset_b,
-		});
-		let share_token = XYK::share_token(pair_account);
-
-		assert_eq!(Currency::free_balance(asset_a, &pair_account), 1004000000001);
-		assert_eq!(Currency::free_balance(asset_b, &pair_account), 100400000);
-		assert_eq!(Currency::free_balance(asset_b, &user), 999999899600000);
-		assert_eq!(Currency::free_balance(share_token, &user), 1004000000000);
-		assert_eq!(XYK::total_liquidity(&pair_account), 1004000000000);
-
-		assert_ok!(XYK::add_liquidity(
-			Origin::signed(BOB),
-			asset_b,
-			asset_a,
-			1_000_000,
-			1_000_000_000_000
-		));
-
-		assert_eq!(Currency::free_balance(asset_a, &pair_account), 1014000000002);
-		assert_eq!(Currency::free_balance(asset_b, &pair_account), 101400000);
-		assert_eq!(Currency::free_balance(asset_b, &user), 999999899600000);
-		assert_eq!(Currency::free_balance(asset_b, &BOB), 999999999000000);
-		assert_eq!(Currency::free_balance(share_token, &user), 1004000000000);
-		assert_eq!(Currency::free_balance(share_token, &BOB), 10000000000);
-		assert_eq!(XYK::total_liquidity(&pair_account), 1014000000000);
-
-		expect_events(vec![
-			Event::PoolCreated {
-				who: ALICE,
-				asset_a: asset_b,
-				asset_b: asset_a,
-				initial_shares_amount: 1000000000000,
-				share_token,
-				pool: pair_account,
-			}
-			.into(),
-			Event::LiquidityAdded {
-				who: ALICE,
-				asset_a: asset_b,
-				asset_b: asset_a,
-				amount_a: 400000,
-				amount_b: 4000000001,
-			}
-			.into(),
-			orml_tokens::Event::Endowed {
-				currency_id: 0,
-				who: 2,
-				amount: 10000000000,
-			}
-			.into(),
-			Event::LiquidityAdded {
-				who: BOB,
-				asset_a: asset_b,
-				asset_b: asset_a,
-				amount_a: 1000000,
-				amount_b: 10000000001,
-			}
-			.into(),
-		]);
-	});
-}
-
-#[test]
-fn remove_liquidity_should_work() {
-	new_test_ext().execute_with(|| {
-		let user = ALICE;
-		let asset_a = HDX;
-		let asset_b = DOT;
-
-		assert_ok!(XYK::create_pool(
-			Origin::signed(user),
-			asset_a,
-			asset_b,
-			100_000_000,
-			Price::from(10_000)
-		));
-
-		let pair_account = XYK::get_pair_id(AssetPair {
-			asset_in: asset_a,
-			asset_out: asset_b,
-		});
-		let share_token = XYK::share_token(pair_account);
-
-		assert_eq!(Currency::free_balance(share_token, &user), 100000000);
-		assert_eq!(Currency::free_balance(asset_a, &user), 999999900000000);
-		assert_eq!(Currency::free_balance(asset_a, &pair_account), 100000000);
-		assert_eq!(Currency::free_balance(asset_b, &pair_account), 1000000000000);
-
-		assert_ok!(XYK::remove_liquidity(Origin::signed(user), asset_a, asset_b, 355_000));
-
-		assert_eq!(Currency::free_balance(asset_b, &pair_account), 996450000000);
-		assert_eq!(Currency::free_balance(asset_a, &user), 999999900355000);
-
-		assert_eq!(Currency::free_balance(share_token, &user), 99645000);
-		assert_eq!(XYK::total_liquidity(&pair_account), 99645000);
-
-		expect_events(vec![
-			Event::PoolCreated {
-				who: ALICE,
-				asset_a,
-				asset_b,
-				initial_shares_amount: 100000000,
-				share_token,
-				pool: pair_account,
-			}
-			.into(),
-			Event::LiquidityRemoved {
-				who: ALICE,
-				asset_a,
-				asset_b,
-				shares: 355_000,
-			}
-			.into(),
-		]);
-	});
-}
-
-#[test]
-fn remove_liquidity_without_shares_should_not_work() {
-	new_test_ext().execute_with(|| {
-		let user = ALICE;
-		let asset_a = HDX;
-		let asset_b = DOT;
-
-		assert_ok!(XYK::create_pool(
-			Origin::signed(user),
-			asset_a,
-			asset_b,
-			100_000_000,
-			Price::from(1)
-		));
-
-		let pair_account = XYK::get_pair_id(AssetPair {
-			asset_in: asset_a,
-			asset_out: asset_b,
-		});
-		let share_token = XYK::share_token(pair_account);
-		let shares = Currency::free_balance(share_token, &user);
-
-		assert_ok!(Currency::transfer(Origin::signed(ALICE), BOB, share_token, shares));
-
-		assert_noop!(
-			XYK::remove_liquidity(Origin::signed(user), asset_a, asset_b, 355_000),
-			Error::<Test>::InsufficientAssetBalance
-		);
-
-		expect_events(vec![
-			Event::PoolCreated {
-				who: ALICE,
-				asset_a,
-				asset_b,
-				initial_shares_amount: 100000000,
-				share_token,
-				pool: pair_account,
-			}
-			.into(),
-			orml_tokens::Event::Endowed {
-				currency_id: share_token,
-				who: BOB,
-				amount: shares,
-			}
-			.into(),
-			orml_tokens::Event::Transfer {
-				currency_id: share_token,
-				from: ALICE,
-				to: BOB,
-				amount: shares,
-			}
-			.into(),
-		]);
-	});
-}
-
-// events in the following test do not occur during standard chain operation
-#[test]
-fn remove_liquidity_from_reduced_pool_should_not_work() {
-	new_test_ext().execute_with(|| {
-		let user = ALICE;
-		let asset_a = HDX;
-		let asset_b = DOT;
-
-		assert_ok!(XYK::create_pool(
-			Origin::signed(user),
-			asset_a,
-			asset_b,
-			100_000_000,
-			Price::from(1)
-		));
-
-		let pair_account = XYK::get_pair_id(AssetPair {
-			asset_in: asset_a,
-			asset_out: asset_b,
-		});
-
-		// remove some amount from the pool
-		assert_ok!(Currency::transfer(
-			Origin::signed(pair_account),
-			BOB,
-			asset_a,
-			90_000_000
-		));
-
-		assert_noop!(
-			XYK::remove_liquidity(Origin::signed(user), asset_a, asset_b, 200_000_000),
-			Error::<Test>::InsufficientLiquidity
-		);
-
-		// return it back to the pool
-		assert_ok!(Currency::transfer(
-			Origin::signed(BOB),
-			pair_account,
-			asset_a,
-			90_000_000
-		));
-		// do it again with asset_b
-		assert_ok!(Currency::transfer(
-			Origin::signed(pair_account),
-			BOB,
-			asset_b,
-			90_000_000
-		));
-
-		assert_noop!(
-			XYK::remove_liquidity(Origin::signed(user), asset_a, asset_b, 200_000_000),
-			Error::<Test>::InsufficientLiquidity
-		);
-
-		let pair_account = XYK::get_pair_id(AssetPair {
-			asset_in: asset_a,
-			asset_out: asset_b,
-		});
-		let share_token = XYK::share_token(pair_account);
-
-		expect_events(vec![
-			Event::PoolCreated {
-				who: ALICE,
-				asset_a,
-				asset_b,
-				initial_shares_amount: 100000000,
-				share_token,
-				pool: pair_account,
-			}
-			.into(),
-			orml_tokens::Event::Transfer {
-				currency_id: asset_a,
-				from: pair_account,
-				to: BOB,
-				amount: 90_000_000,
-			}
-			.into(),
-			orml_tokens::Event::Transfer {
-				currency_id: asset_a,
-				from: BOB,
-				to: pair_account,
-				amount: 90_000_000,
-			}
-			.into(),
-			orml_tokens::Event::Transfer {
-				currency_id: asset_b,
-				from: pair_account,
-				to: BOB,
-				amount: 90_000_000,
-			}
-			.into(),
-		]);
-	});
-}
-
-#[test]
-fn add_liquidity_more_than_owner_should_not_work() {
-	new_test_ext().execute_with(|| {
-		assert_ok!(XYK::create_pool(
-			Origin::signed(ALICE),
-			HDX,
-			ACA,
-			200_000_000,
-			Price::from(3000000)
-		));
-
-		assert_eq!(Currency::free_balance(ACA, &ALICE), 400000000000000);
-
-		assert_noop!(
-			XYK::add_liquidity(Origin::signed(ALICE), HDX, ACA, 200_000_000_000_000_000, 600_000_000),
-			Error::<Test>::InsufficientAssetBalance
-		);
-
-		assert_noop!(
-			XYK::add_liquidity(Origin::signed(ALICE), HDX, ACA, 600_000_000, 200_000_000_000_000_000),
-			Error::<Test>::InsufficientAssetBalance
-		);
-	});
-}
-
-#[test]
-fn add_insufficient_liquidity_should_not_work() {
-	new_test_ext().execute_with(|| {
-		assert_ok!(XYK::create_pool(
-			Origin::signed(ALICE),
-			HDX,
-			ACA,
-			1000,
-			Price::from_float(1.5)
-		));
-
-		assert_noop!(
-			XYK::add_liquidity(Origin::signed(ALICE), HDX, ACA, 0, 0),
-			Error::<Test>::InsufficientTradingAmount
-		);
-
-		assert_noop!(
-			XYK::add_liquidity(Origin::signed(ALICE), HDX, ACA, 1000, 0),
-			Error::<Test>::ZeroLiquidity
-		);
-
-		assert_noop!(
-			XYK::add_liquidity(Origin::signed(BOB), ACA, HDX, 1000, 2000),
-			Error::<Test>::InsufficientLiquidity
-		);
-	});
-}
-
-#[test]
-fn add_liquidity_exceeding_max_limit_should_not_work() {
-	new_test_ext().execute_with(|| {
-		assert_ok!(XYK::create_pool(
-			Origin::signed(ALICE),
-			HDX,
-			ACA,
-			100_000_000_000_000,
-			Price::from(1)
-		));
-
-		assert_noop!(
-			XYK::add_liquidity(Origin::signed(ALICE), HDX, ACA, 10_000_000, 1_000_000),
-			Error::<Test>::AssetAmountExceededLimit
-		);
-	});
-}
-#[test]
-fn remove_liquidity_should_respect_min_pool_limit() {
-	new_test_ext().execute_with(|| {
-		assert_ok!(XYK::create_pool(
-			Origin::signed(ALICE),
-			HDX,
-			ACA,
-			1000,
-			Price::from_float(1.5)
-		));
-
-		assert_ok!(XYK::add_liquidity(Origin::signed(BOB), ACA, HDX, 2000, 2000));
-
-		assert_noop!(
-			XYK::remove_liquidity(Origin::signed(BOB), ACA, HDX, 500),
-			Error::<Test>::InsufficientLiquidity
-		);
-	});
-}
-
-#[test]
-fn remove_zero_liquidity_should_not_work() {
-	new_test_ext().execute_with(|| {
-		assert_noop!(
-			XYK::remove_liquidity(Origin::signed(ALICE), HDX, ACA, 0),
-			Error::<Test>::ZeroLiquidity
-		);
-	});
-}
+use primitives::asset::AssetPair;
+use primitives::Price;
 
 #[test]
 fn sell_test() {
@@ -1440,46 +790,6 @@ fn single_buy_with_discount_should_work() {
 }
 
 #[test]
-fn create_pool_with_insufficient_liquidity_should_not_work() {
-	new_test_ext().execute_with(|| {
-		assert_noop!(
-			XYK::create_pool(Origin::signed(ALICE), ACA, HDX, 500, Price::from(3200)),
-			Error::<Test>::InsufficientLiquidity
-		);
-
-		assert_noop!(
-			XYK::create_pool(Origin::signed(ALICE), ACA, HDX, 5000, Price::from_float(0.1f64)),
-			Error::<Test>::InsufficientLiquidity
-		);
-
-		assert_noop!(
-			XYK::create_pool(Origin::signed(ALICE), ACA, HDX, 1000, Price::from(0)),
-			Error::<Test>::ZeroInitialPrice
-		);
-	});
-}
-
-#[test]
-fn add_liquidity_to_non_existing_pool_should_not_work() {
-	new_test_ext().execute_with(|| {
-		assert_noop!(
-			XYK::add_liquidity(Origin::signed(ALICE), HDX, ACA, 200_000_000_000_000_000, 600_000_000),
-			Error::<Test>::TokenPoolNotFound
-		);
-	});
-}
-
-#[test]
-fn remove_zero_liquidity_from_non_existing_pool_should_not_work() {
-	new_test_ext().execute_with(|| {
-		assert_noop!(
-			XYK::remove_liquidity(Origin::signed(ALICE), HDX, ACA, 100),
-			Error::<Test>::TokenPoolNotFound
-		);
-	});
-}
-
-#[test]
 fn sell_with_non_existing_pool_should_not_work() {
 	new_test_ext().execute_with(|| {
 		assert_noop!(
@@ -1532,83 +842,6 @@ fn discount_buy_with_no_native_pool_should_not_work() {
 			XYK::buy(Origin::signed(ALICE), ACA, DOT, 1000, 1_000_000_000, true),
 			Error::<Test>::CannotApplyDiscount
 		);
-	});
-}
-
-#[test]
-fn create_pool_small_fixed_point_amount_should_work() {
-	new_test_ext().execute_with(|| {
-		let asset_a = HDX;
-		let asset_b = ACA;
-
-		assert_ok!(XYK::create_pool(
-			Origin::signed(ALICE),
-			asset_a,
-			asset_b,
-			100_000_000_000_000,
-			Price::from_float(0.00001)
-		));
-
-		let pair_account = XYK::get_pair_id(AssetPair {
-			asset_in: asset_a,
-			asset_out: asset_b,
-		});
-		let share_token = XYK::share_token(pair_account);
-
-		assert_eq!(Currency::free_balance(asset_a, &pair_account), 100000000000000);
-		assert_eq!(Currency::free_balance(asset_b, &pair_account), 1000000000);
-		assert_eq!(Currency::free_balance(asset_a, &ALICE), 900000000000000);
-		assert_eq!(Currency::free_balance(asset_b, &ALICE), 999999000000000);
-		assert_eq!(Currency::free_balance(share_token, &ALICE), 100000000000000);
-		assert_eq!(XYK::total_liquidity(&pair_account), 100000000000000);
-
-		expect_events(vec![Event::PoolCreated {
-			who: ALICE,
-			asset_a,
-			asset_b,
-			initial_shares_amount: 100000000000000,
-			share_token,
-			pool: pair_account,
-		}
-		.into()]);
-	});
-}
-
-#[test]
-fn create_pool_fixed_point_amount_should_work() {
-	new_test_ext().execute_with(|| {
-		let asset_a = HDX;
-		let asset_b = ACA;
-		assert_ok!(XYK::create_pool(
-			Origin::signed(ALICE),
-			asset_a,
-			asset_b,
-			100_000_000_000,
-			Price::from_float(4560.234543)
-		));
-
-		let pair_account = XYK::get_pair_id(AssetPair {
-			asset_in: asset_a,
-			asset_out: asset_b,
-		});
-		let share_token = XYK::share_token(pair_account);
-
-		assert_eq!(Currency::free_balance(asset_a, &pair_account), 100000000000);
-		assert_eq!(Currency::free_balance(asset_b, &pair_account), 456023454299999);
-		assert_eq!(Currency::free_balance(asset_a, &ALICE), 999900000000000);
-		assert_eq!(Currency::free_balance(asset_b, &ALICE), 543976545700001);
-		assert_eq!(Currency::free_balance(share_token, &ALICE), 100000000000);
-		assert_eq!(XYK::total_liquidity(&pair_account), 100000000000);
-
-		expect_events(vec![Event::PoolCreated {
-			who: ALICE,
-			asset_a,
-			asset_b,
-			initial_shares_amount: 100000000000,
-			share_token,
-			pool: pair_account,
-		}
-		.into()]);
 	});
 }
 
@@ -1743,123 +976,6 @@ fn money_in_money_out_should_leave_the_same_balance_for_both_accounts() {
 
 		assert!(!XYK::exists(asset_pair));
 	});
-}
-
-#[test]
-fn destroy_pool_on_remove_liquidity_and_recreate_should_work() {
-	new_test_ext().execute_with(|| {
-		let user = ALICE;
-		let asset_a = HDX;
-		let asset_b = DOT;
-
-		assert_ok!(XYK::create_pool(
-			Origin::signed(user),
-			asset_a,
-			asset_b,
-			100_000_000,
-			Price::from(10_000)
-		));
-
-		let asset_pair = AssetPair {
-			asset_in: asset_a,
-			asset_out: asset_b,
-		};
-
-		let pair_account = XYK::get_pair_id(asset_pair);
-		let share_token = XYK::share_token(pair_account);
-
-		assert!(XYK::exists(asset_pair));
-
-		assert_ok!(XYK::remove_liquidity(
-			Origin::signed(user),
-			asset_a,
-			asset_b,
-			100_000_000
-		));
-
-		assert_eq!(XYK::total_liquidity(&pair_account), 0);
-
-		assert!(!XYK::exists(asset_pair));
-
-		// It should be possible to recreate the pool again
-
-		assert_ok!(XYK::create_pool(
-			Origin::signed(user),
-			asset_a,
-			asset_b,
-			100_000_000,
-			Price::from(10_000)
-		));
-
-		expect_events(vec![
-			Event::PoolCreated {
-				who: user,
-				asset_a,
-				asset_b,
-				initial_shares_amount: 100_000_000,
-				share_token,
-				pool: pair_account,
-			}
-			.into(),
-			frame_system::Event::KilledAccount { account: pair_account }.into(),
-			Event::LiquidityRemoved {
-				who: user,
-				asset_a,
-				asset_b,
-				shares: 100_000_000,
-			}
-			.into(),
-			Event::PoolDestroyed {
-				who: user,
-				asset_a,
-				asset_b,
-				share_token,
-				pool: pair_account,
-			}
-			.into(),
-			frame_system::Event::NewAccount { account: pair_account }.into(),
-			orml_tokens::Event::Endowed {
-				currency_id: asset_a,
-				who: pair_account,
-				amount: 100000000,
-			}
-			.into(),
-			orml_tokens::Event::Endowed {
-				currency_id: asset_b,
-				who: pair_account,
-				amount: 1000000000000,
-			}
-			.into(),
-			orml_tokens::Event::Endowed {
-				currency_id: 0,
-				who: 1,
-				amount: 100000000,
-			}
-			.into(),
-			Event::PoolCreated {
-				who: user,
-				asset_a,
-				asset_b,
-				initial_shares_amount: 100_000_000,
-				share_token,
-				pool: pair_account,
-			}
-			.into(),
-		]);
-	});
-}
-
-#[test]
-fn create_pool_with_same_assets_should_not_be_allowed() {
-	new_test_ext().execute_with(|| {
-		let user = ALICE;
-		let asset_a = HDX;
-
-		assert_noop!(
-			XYK::create_pool(Origin::signed(user), asset_a, asset_a, 100_000_000, Price::from(10_000)),
-			Error::<Test>::CannotCreatePoolWithSameAssets
-		);
-	})
 }
 
 #[test]
@@ -2043,50 +1159,6 @@ fn single_sell_more_than_ratio_in_should_not_work() {
 }
 
 #[test]
-fn test_calculate_out_given_in() {
-	ExtBuilder::default().build().execute_with(|| {
-		let in_reserve: Balance = 10000000000000;
-		let out_reserve: Balance = 100000;
-		let in_amount: Balance = 100000000000;
-		let result = hydra_dx_math::xyk::calculate_out_given_in(in_reserve, out_reserve, in_amount);
-		assert_eq!(result, Ok(990));
-	});
-}
-
-#[test]
-fn test_calculate_out_given_in_invalid() {
-	ExtBuilder::default().build().execute_with(|| {
-		let in_reserve: Balance = 0;
-		let out_reserve: Balance = 1000;
-		let in_amount: Balance = 0;
-		let result = hydra_dx_math::xyk::calculate_out_given_in(in_reserve, out_reserve, in_amount);
-		assert_eq!(result, Ok(0));
-	});
-}
-
-#[test]
-fn test_calculate_in_given_out_insufficient_pool_balance() {
-	ExtBuilder::default().build().execute_with(|| {
-		let in_reserve: Balance = 10000000000000;
-		let out_reserve: Balance = 100000;
-		let out_amount: Balance = 100000000000;
-		let result = hydra_dx_math::xyk::calculate_in_given_out(out_reserve, in_reserve, out_amount);
-		assert_eq!(result, Err(MathError::InsufficientOutReserve));
-	});
-}
-
-#[test]
-fn test_calculate_in_given_out() {
-	ExtBuilder::default().build().execute_with(|| {
-		let in_reserve: Balance = 10000000000000;
-		let out_reserve: Balance = 10000000;
-		let out_amount: Balance = 1000000;
-		let result = hydra_dx_math::xyk::calculate_in_given_out(out_reserve, in_reserve, out_amount);
-		assert_eq!(result, Ok(1111111111112));
-	});
-}
-
-#[test]
 fn sell_with_low_amount_should_not_work() {
 	new_test_ext().execute_with(|| {
 		assert_noop!(
@@ -2122,175 +1194,4 @@ fn buy_with_excesive_amount_should_not_work() {
 			Error::<Test>::InsufficientPoolAssetBalance
 		);
 	});
-}
-
-#[test]
-fn fee_calculation() {
-	ExtBuilder::default().build().execute_with(|| {
-		assert_eq!(XYK::calculate_discounted_fee(10000), Ok(7));
-		assert_eq!(XYK::calculate_discounted_fee(100000), Ok(70));
-		assert_eq!(XYK::calculate_discounted_fee(100000), Ok(70));
-
-		assert_eq!(XYK::calculate_fee(100000), Ok(200));
-		assert_eq!(XYK::calculate_fee(10000), Ok(20));
-	});
-	ExtBuilder::default()
-		.with_exchange_fee((10, 1000))
-		.build()
-		.execute_with(|| {
-			assert_eq!(XYK::calculate_fee(100000), Ok(1000));
-			assert_eq!(XYK::calculate_fee(10000), Ok(100));
-		});
-
-	ExtBuilder::default()
-		.with_exchange_fee((10, 0))
-		.build()
-		.execute_with(|| {
-			assert_eq!(XYK::calculate_fee(100000), Ok(0));
-		});
-
-	ExtBuilder::default()
-		.with_exchange_fee((10, 1))
-		.build()
-		.execute_with(|| {
-			assert_noop!(XYK::calculate_fee(u128::MAX), Error::<Test>::FeeAmountInvalid);
-		});
-}
-
-#[test]
-fn can_create_pool_should_work() {
-	new_test_ext().execute_with(|| {
-		let asset_a = 10u32;
-		let asset_b = 10u32;
-		assert_noop!(
-			XYK::create_pool(
-				Origin::signed(ALICE),
-				asset_a,
-				asset_b,
-				100_000_000_000_000,
-				Price::from(10)
-			),
-			Error::<Test>::CannotCreatePool
-		);
-	});
-}
-
-#[test]
-fn get_fee_should_work() {
-	new_test_ext().execute_with(|| {
-		assert_ok!(XYK::create_pool(
-			Origin::signed(ALICE),
-			HDX,
-			DOT,
-			1_000_000_000,
-			Price::from(2)
-		),);
-
-		// existing pool
-		let fee = XYK::get_fee(&HDX_DOT_POOL_ID);
-		assert_eq!(fee, (2, 1_000));
-		// non existing pool
-		let fee = XYK::get_fee(&1_234);
-		assert_eq!(fee, (2, 1_000));
-	});
-}
-
-#[test]
-fn share_ratio_calculations_are_correct() {
-	ExtBuilder::default()
-		.with_exchange_fee((0, 0))
-		.build()
-		.execute_with(|| {
-			let asset_a = HDX;
-			let asset_b = DOT;
-
-			assert_ok!(XYK::create_pool(
-				Origin::signed(ALICE),
-				asset_a,
-				asset_b,
-				100 * ONE,
-				Price::from_float(0.6544)
-			));
-
-			assert_eq!(Currency::free_balance(asset_a, &BOB), 1_000 * ONE);
-			assert_eq!(Currency::free_balance(asset_b, &BOB), 1_000 * ONE);
-
-			let balance_a = Currency::free_balance(asset_a, &BOB);
-			let balance_b = Currency::free_balance(asset_b, &BOB);
-
-			let bob_initial_balance = balance_a + balance_b;
-
-			assert_eq!(bob_initial_balance, 2000 * ONE);
-
-			assert_ok!(XYK::add_liquidity(
-				Origin::signed(BOB),
-				asset_b,
-				asset_a,
-				10 * ONE,
-				200 * ONE
-			));
-
-			let pair_account = XYK::get_pair_id(AssetPair {
-				asset_in: asset_a,
-				asset_out: asset_b,
-			});
-			let share_token = XYK::share_token(pair_account);
-
-			let expected_shares = 15_281_173_594_132u128;
-
-			assert_eq!(Currency::free_balance(share_token, &BOB), expected_shares);
-
-			assert_ok!(XYK::sell(
-				Origin::signed(CHARLIE),
-				asset_a,
-				asset_b,
-				10 * ONE,
-				0u128,
-				false,
-			));
-
-			assert_ok!(XYK::remove_liquidity(
-				Origin::signed(BOB),
-				asset_a,
-				asset_b,
-				expected_shares
-			));
-
-			assert_eq!(Currency::free_balance(share_token, &BOB), 0);
-
-			for _ in 0..10 {
-				let balance_a = Currency::free_balance(asset_a, &BOB);
-				let balance_b = Currency::free_balance(asset_b, &BOB);
-
-				let bob_previous_balance = balance_a + balance_b;
-
-				let balance_pool_a = Currency::free_balance(asset_a, &pair_account);
-				let balance_pool_b = Currency::free_balance(asset_a, &pair_account);
-
-				let initial_pool_liquidity = balance_pool_a + balance_pool_b;
-
-				assert_ok!(XYK::add_liquidity(
-					Origin::signed(BOB),
-					asset_b,
-					asset_a,
-					10 * ONE,
-					200 * ONE
-				));
-
-				let shares = Currency::free_balance(share_token, &BOB);
-
-				assert_ok!(XYK::remove_liquidity(Origin::signed(BOB), asset_a, asset_b, shares));
-				let balance_a = Currency::free_balance(asset_a, &BOB);
-				let balance_b = Currency::free_balance(asset_b, &BOB);
-				let bob_new_balance = balance_a + balance_b;
-
-				let balance_pool_a = Currency::free_balance(asset_a, &pair_account);
-				let balance_pool_b = Currency::free_balance(asset_a, &pair_account);
-
-				let total_pool_liquidity = balance_pool_a + balance_pool_b;
-
-				assert!(bob_new_balance <= bob_previous_balance);
-				assert!(initial_pool_liquidity <= total_pool_liquidity);
-			}
-		});
 }
